+++ tools/web3-plugin-example/src/transaction_middleware.ts
@@ -14,19 +14,25 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { DataFormat } from 'web3-types';
-import { Web3Subscription } from '../../../src';
 
-export class ExampleSubscription extends Web3Subscription<
-	{ data: string },
-	{ param1: string },
-	{ eth_subscribe: (newHeads: string) => void }
-> {
+import { TransactionMiddleware, TransactionMiddlewareData } from "web3-eth";
+
+// Sample Transaction Middleware 
+export class Web3TransactionMiddleware implements TransactionMiddleware {
+
 	// eslint-disable-next-line class-methods-use-this
-	protected _buildSubscriptionParams() {
-		return ['newHeads'];
+	public async processTransaction(transaction: TransactionMiddlewareData, 
+		_options?: { [key: string]: unknown } | undefined): 
+		
+		Promise<TransactionMiddlewareData> {
+		
+		// eslint-disable-next-line prefer-const
+		let txObj = { ...transaction };
+
+		// Add your logic here for transaction modification 
+		txObj.data = '0x123';
+
+		return Promise.resolve(txObj);
 	}
-	public getReturnFormat(): DataFormat {
-		return this.returnFormat;
-	}
+
 }