+++ tools/web3-plugin-example/CHANGELOG.md
@@ -89,3 +89,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Dependencies updated
 
 ## [Unreleased]
+
+### Added
+
+Transaction middleware (#7088)